Make a separate service to do template transformations
